requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.8.5
Pillow==9.5.0
geopy==2.4.1
hijri-converter==2.3.2